fix(2521): split messageText helpers + register 3 env vars
Two CI-only failures from PR #2578 run 3:

────────────────────────────────────────────────────────────────────────
TUI Bun loader — split tiny helpers off messages.ts
────────────────────────────────────────────────────────────────────────

The forward-ref reorder in commit 5739687 didn't help — Linux CI Bun
1.3.x continues to drop the `isEmptyMessageText` export from the
5,331-line `tui/src/utils/messages.ts`. macOS Bun 1.3.12 still loads
the same file cleanly. The Bun loader's "Unhandled error between
tests" surfaces as

    SyntaxError: Export named 'isEmptyMessageText' not found in module
    '/.../tui/src/utils/messages.ts'

…on every CI run, even when the symbol is verifiably exported.
The error pointer (`1 | })\n2 | {`) suggests a parser stall on a
completely unrelated line of the 5K-line file when Linux Bun
evaluates the module — but the macOS build accepts it.

Lowest-risk structural fix: pull the two pure helpers into a tiny
new module `tui/src/utils/messageText.ts` (32 lines) and re-export
them from messages.ts so existing call sites still resolve. Files
that previously imported these helpers from messages.ts now go to
messageText.ts directly:

  tui/src/utils/markdown.ts
  tui/src/components/Markdown.tsx

This means thinking-delta-render.test.tsx's transitive load chain
(test → AssistantThinkingMessage → Markdown → markdown.ts) now
hits the small file instead of messages.ts and the loader edge
case is bypassed. Other consumers (AssistantTextMessage,
MessageSelector, copy.tsx, messageActions) still use messages.ts
because they need other unrelated symbols from it, but the failing
test no longer does.

────────────────────────────────────────────────────────────────────────
Env registry — register 3 KOSMOS_* vars introduced in this PR
────────────────────────────────────────────────────────────────────────

`uv run python scripts/audit-env-registry.py` (Spec 026 contract,
gate in `Test (Python 3.12 / 3.13)`) flagged 3 unregistered
references:

  KOSMOS_AVAILABLE_ADAPTERS_TOP_K   — src/kosmos/ipc/stdio.py:767
  KOSMOS_LLM_STREAM_CHUNK_MAX_CHARS — src/kosmos/llm/client.py:70/81
  KOSMOS_LLM_STREAM_PACE_MS         — src/kosmos/llm/client.py:71/83

Added rows to `docs/configuration.md` with description, default,
allowed values, code-side anchor, and the originating spec.

  Verdict: drift → registry rows 84 → 87 → drift cleared.

────────────────────────────────────────────────────────────────────────
Verification
────────────────────────────────────────────────────────────────────────

  uv run python scripts/audit-env-registry.py: in_code_not_in_registry: []
  bun typecheck: clean
  bun test: 974 pass / 0 fail
  bun test tests/ipc/thinking-delta-render.test.tsx: 4 pass

  Linux CI to confirm the Bun loader is now unblocked.
