+++ docs/configuration.md
@@ -59,6 +59,9 @@ Column definitions:
 | `KOSMOS_PROMPTS_DIR` | No | `<repo_root>/prompts` | Absolute or relative directory path | `kosmos.tools.verify_canonical_map._resolve_prompts_dir` (Epic ζ #2297 — overrides the path that `verify_canonical_map.py` parses to source the canonical 10-row `tool_id ↔ family_hint` mapping from `<verify_families>` in `system_v1.md`; defaults to walking up from the module's parent directories to find `prompts/`) | Epic #2297 |
 | `KOSMOS_PERMISSION_TIMEOUT_SECONDS` | No | `60` | Float > 0 (seconds) | `kosmos.ipc.stdio` (Spec 1978 T045 — permission_request → permission_response wait; D2 invariant default-deny on timeout) | Spec 1978 |
 | `KOSMOS_K_EXAONE_THINKING` | No | `true` | `true` \| `false` (case-insensitive; `1`/`yes` also accepted) | `kosmos.llm.client._build_payload` (Epic #2077 / Spec 2521 — opts in to K-EXAONE-236B-A23B's chain-of-thought channel via `chat_template_kwargs.enable_thinking`; default `true` per K-EXAONE model card and τ²-Bench measurement conditions; set `false` for sub-second first-token latency at the cost of losing the citizen-visible `∴ Thinking` glyph) | Epic #2077 / Spec 2521 |
+| `KOSMOS_AVAILABLE_ADAPTERS_TOP_K` | No | `5` | Integer >= 1 | `kosmos.ipc.stdio._build_available_adapters_suffix` (Spec 2521 — bounds how many BM25 candidates are emitted into the dynamic `<available_adapters>` system-prompt suffix per citizen turn; lower for prompt-cache friendliness, higher for broader recall) | Spec 2521 |
+| `KOSMOS_LLM_STREAM_CHUNK_MAX_CHARS` | No | `999` | Integer >= 1 (chars) | `kosmos.llm.client._pace_text_chunk` (Spec 2521 — when set <999, splits provider deltas into sub-chunks for headless / no-Ink callers that want server-side cadence; default `999` is effectively "no extra splitting" because Ink frontend typewriter handles the in-TUI cadence) | Spec 2521 |
+| `KOSMOS_LLM_STREAM_PACE_MS` | No | `0` | Float >= 0 (milliseconds) | `kosmos.llm.client._pace_text_chunk` (Spec 2521 — sleep between sub-chunk emissions for headless callers; default `0` disables backend pacing because Ink's `FRAME_INTERVAL_MS=4` throttle relax handles the cadence inside the TUI) | Spec 2521 |
 | `KOSMOS_LOOKUP_TOPK` | No | `5` | Integer [1, 20] | `kosmos.settings.KosmosSettings.lookup_topk` | This doc |
 | `KOSMOS_NMC_FRESHNESS_MINUTES` | No | `30` | Integer [1, 1440] (minutes) | `kosmos.settings.KosmosSettings.nmc_freshness_minutes` | Epic #507 |
 | `KOSMOS_RETRIEVAL_BACKEND` | No | `bm25` | `bm25` \| `dense` \| `hybrid` | `kosmos.tools.retrieval.backend.build_retriever_from_env` | Epic #585 |
