+++ tui/src/utils/markdown.ts
@@ -8,7 +8,7 @@ import { supportsHyperlinks } from '../ink/supports-hyperlinks.js'
 import type { CliHighlight } from './cliHighlight.js'
 import { logForDebugging } from './debug.js'
 import { createHyperlink } from './hyperlink.js'
-import { stripPromptXMLTags } from './messages.js'
+import { stripPromptXMLTags } from './messageText.js'
 import type { ThemeName } from './theme.js'
 
 // Use \n unconditionally — os.EOL is \r\n on Windows, and the extra \r
